Fix PostgreSQL installation
Missing target version was omitting postgresql 16 from the image and
copying /usr/pgsql-* copies the dir contents which is not what we want.

Currently hardcoding the list of versions in the dockerfile. A proper
fix might be to just copy whole /usr into a temp location and use a
shell script to pick out stuff we need, remove the temp afterwards.